Deploy site to new AWS bucket and domain
I was able to get DNS zone editing access to the `austintexas.io` domain, so I created a `developer-guides` subdomain to host this site. There were a couple of components involved in that migration:

- AWS Route 53 zone editing, through the Innovation Office AWS organization: the `austintexas.io` domain is managed through inno's AWS account, so I had to use that aws console to create the zone change. I added the `developer-guides` subdomain as an `A Record` and pointed it to a corresponding bucket per instructions at http://docs.aws.amazon.com/Route53/latest/DeveloperGuide/resource-record-sets-values-alias.html. The important thing to note here is that **the bucket name had to match the entire URL**. That's why I had to create a bucket called `developer-guides.austintexas.io`. Dirty, I know. It is what it is.
- Updated the `baseurl` setting in `_config.yml` to reference the new pretty domain.

Longer-term I don't think it makes sense to have separate AWS accounts for different city orgs, specifically for situations like this. It would have been way easier for me to update the DNS setting if the S3 bucket was within the same account as the Route 53 domain management. There are ways to create access privileges using AWS to make sure that only the right folks have access to the `austintexas.io` zone. Fortunately I was sitting with someone who had the permissions-level to invite me to the DTI organization on AWS so I could make the changes myself. Now I'm part of 3 separate AWS orgs (ACCD, DTI, CTM), which is pretty unsustainable.