Remove "baseurl" property from Jekyll config file
I had initially put this in because I was having issues getting the deployed site to both pull in the assets _and_ have the nav links point to the right pages. What I didn't realize is that this creates a conflict when running locally as the Jekyll server appends that property to the end of the localhost path. My first step to troubleshoot this was to remove the baseurl property entirely and re-deploy to AWS and, lo and behold, everything was working fine. I'm not sure what the original issue was but it all appears to be resolved by just removing the line.