Fix multi-environment deployment config in Travis yml
I noticed that updates to `master` weren't getting deployed out to S3, and that the Travis build log included a message saying:

```
Skipping a deployment with the s3 provider because this branch is not permitted
```

After a bunch of investigation I realized the issue came as a result of declaring 2 separate "deploy:" settings in `.travis.yml`. The correct way of handling that is reflected in this diff, where each _provider_ is listed separately within the single "deploy:" declaration.

This still leads to some undesirable behavior, however.

**When updating master** (which is the first branch specified in the deployment rules) the build log will show the expected message around the successful deploy.
**When updating staging** (which is the second branch specified in the deployment rules), the build log will _first_ show that it's skipping deployment because the branch is not permitted (while evaluating for `master`), and THEN, as it hits the second deployment rule, will show that the deployment succeeded as it evaluates for `staging`. I looked into ways to customize or override that log output but was not able to find anything.