Add Knackpy to library of microservices
Though Knackpy isn't technically a _microservice_, it is in-line with the broader CoA/CTM vision of "Shared IT" so I thought it was worth it to list it in the library. This got me thinking that maybe the "Microservices" top-level section should instead be called something like "Shared Technology" and use it to organize the paid vendor solutions, open-source tools, custom-built software, and microservices that are in use across departments at the city.